created brainbotmodal, added ai button
