+++ frontend/app/_layout.tsx
@@ -10,7 +10,6 @@ export default function RootLayout() {
     <Stack screenOptions={{ headerShown: false }}>
       <Stack.Screen name="(auth)" options={{ headerShown: false }} />
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-      {/* <Stack.Screen name="(tabs)/chatbot" options={{ headerShown: false }} /> */}
     </Stack>
   );
 }
